requirements.txt
flask
tensorflow
line-bot-sdk
numpy
opencv-python
pytesseract
